Moves images from /shared/images to /images in folder tree
