+++ src/signals/incident-management/definitions/listIcons.tsx
@@ -1,4 +1,4 @@
-import IconPriorityHigh from '../../../shared/images/icon-urgentie-hoog.svg'
+import IconPriorityHigh from '../../../images/icon-urgentie-hoog.svg'
 
 export type Icon = 'PriorityHigh'
